refactoring :: RecruitmentsDomainInterface Transfer 나눔
