+++ Projects/Domain/RecruitmentsDomain/Interface/DTO/Response/RecruitmentDetailResponseDTO.swift
@@ -66,66 +66,3 @@ public struct RecruitmentDetailResponseDTO: Decodable {
         case etc
     }
 }
-
-public struct AreaResponseDTO: Decodable {
-    public let id: Int
-    public let job: String
-    public let tech: [String]
-    public let hiring: Int
-    public let majorTask: String
-
-    public init(
-        id: Int,
-        job: String,
-        tech: [String],
-        hiring: Int,
-        majorTask: String
-    ) {
-        self.id = id
-        self.job = job
-        self.tech = tech
-        self.hiring = hiring
-        self.majorTask = majorTask
-    }
-
-    enum CodingKeys: String, CodingKey {
-        case id, job, tech, hiring
-        case majorTask = "major_task"
-    }
-}
-
-public extension RecruitmentDetailResponseDTO {
-    func toDomain() -> RecruitmentDetailEntity {
-        RecruitmentDetailEntity(
-            companyID: companyID,
-            areas: areas.map { $0.toDomain() },
-            preferentialTreatment: preferentialTreatment ?? "없음",
-            requiredGrade: requiredGrade,
-            workHours: String(workHours),
-            requiredLicenses: requiredLicenses?.joined(separator: ", ") ?? "없음",
-            hiringProgress: hiringProgress.enumerated().map { (index, value) in
-                "\(index + 1). \(value.localizedString())"
-            }.joined(separator: "\n"),
-            trainPay: String(trainPay),
-            pay: pay,
-            benefits: benefits ?? "없음",
-            military: military,
-            submitDocument: submitDocument,
-            startDate: startDate,
-            endDate: endDate,
-            etc: etc ?? "없음"
-        )
-    }
-}
-
-public extension AreaResponseDTO {
-    func toDomain() -> AreaEntity {
-        AreaEntity(
-            id: String(id),
-            job: job,
-            tech: tech,
-            hiring: hiring,
-            majorTask: majorTask
-        )
-    }
-}
