+++ Projects/Domain/RecruitmentsDomain/Interface/DTO/Trasfer/RecruitmentListTransfer.swift
@@ -1,0 +1,22 @@
+import Foundation
+
+public extension RecruitmentListResponseDTO {
+    func toDomain() -> RecruitmentListEntity {
+        RecruitmentListEntity(recruitments: recruitments.map { $0.toDomain() })
+    }
+}
+
+public extension RecruitmentResponseDTO {
+    func toDomain() -> RecruitmentEntity {
+        RecruitmentEntity(
+            recruitID: recruitID,
+            companyName: companyName,
+            companyProfileURL: companyProfileURL,
+            trainPay: trainPay,
+            military: military,
+            totalHiring: totalHiring,
+            jobCodeList: jobCodeList,
+            bookmarked: bookmarked
+        )
+    }
+}
