+++ Projects/Domain/RecruitmentsDomain/Interface/Entity/AreaEntity.swift
@@ -1,0 +1,23 @@
+import Foundation
+
+public struct AreaEntity: Equatable, Hashable {
+    public let id: String
+    public let job: String
+    public let tech: [String]
+    public let hiring: Int
+    public let majorTask: String
+
+    public init(
+        id: String,
+        job: String,
+        tech: [String],
+        hiring: Int,
+        majorTask: String
+    ) {
+        self.id = id
+        self.job = job
+        self.tech = tech
+        self.hiring = hiring
+        self.majorTask = majorTask
+    }
+}
