+++ Projects/Domain/RecruitmentsDomain/Interface/Entity/RecruitmentDetailEntity.swift
@@ -50,25 +50,3 @@ public struct RecruitmentDetailEntity: Equatable {
         self.etc = etc
     }
 }
-
-public struct AreaEntity: Equatable, Hashable {
-    public let id: String
-    public let job: String
-    public let tech: [String]
-    public let hiring: Int
-    public let majorTask: String
-
-    public init(
-        id: String,
-        job: String,
-        tech: [String],
-        hiring: Int,
-        majorTask: String
-    ) {
-        self.id = id
-        self.job = job
-        self.tech = tech
-        self.hiring = hiring
-        self.majorTask = majorTask
-    }
-}
